requirements.txt
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
torch==1.7.1
torchvision==0.9.0
torchaudio==0.8.0
segmentation-models-pytorch
pip install torch==1.8.0 torchvision==0.9.0 torchaudio==0.8.0
matplotlib